refactor to remove config file (too sparse), move imports to top
